separate vmtelnet functions
